Refactor configuration and redcode modules
